aimd cwnd+=ai/cwnd

+++ datagrump/controller.cc
@@ -6,9 +6,9 @@
 
 using namespace std;
 
-float cwnd = 1;
-float ai = 1;
-float md_factor = 2;
+double cwnd = 1;
+double ai = 1;
+double md_factor = 2;
 unsigned int max_wnd = 22;
 unsigned int last_sequence_number_sent = 0;
 unsigned int last_sequence_number_acked = 0;
